FastAPI/requirements.txt
fastapi==0.133.1
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.1
scikit_learn==1.8.0
SQLAlchemy==2.0.47
uvicorn
pymysql==1.1.2
seaborn==0.13.2
scipy==1.17.0